Split TargetRoots and TargetRootsCalculator to allow TargetRoots to be used more widely without a cycle.
