+++ src/python/pants/bin/goal_runner.py
@@ -8,6 +8,8 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 import logging
 import sys
 
+from twitter.common.collections import OrderedSet
+
 from pants.base.cmd_line_spec_parser import CmdLineSpecParser
 from pants.base.workunit import WorkUnit, WorkUnitLabel
 from pants.bin.engine_initializer import EngineInitializer
@@ -21,11 +23,11 @@ from pants.goal.goal import Goal
 from pants.goal.run_tracker import RunTracker
 from pants.help.help_printer import HelpPrinter
 from pants.init.subprocess import Subprocess
-from pants.init.target_roots import TargetRoots
+from pants.init.target_roots_calculator import TargetRootsCalculator
 from pants.java.nailgun_executor import NailgunProcessGroup
 from pants.option.ranked_value import RankedValue
 from pants.reporting.reporting import Reporting
-from pants.scm.subsystems.changed import Changed
+from pants.scm.subsystems.changed import Changed, ChangedRequest
 from pants.source.source_root import SourceRootConfig
 from pants.task.task import QuietTaskMixin
 from pants.util.filtering import create_filters, wrap_filters
@@ -113,7 +115,7 @@ class GoalRunnerFactory(object):
         include_trace_on_error=self._options.for_global_scope().print_exception_stacktrace
       )
 
-    target_roots = target_roots or TargetRoots.create(
+    target_roots = target_roots or TargetRootsCalculator.create(
       options=self._options,
       build_root=self._root_dir,
       change_calculator=graph_helper.change_calculator
